+++ src/main/java/QCraft/QCraft/Handler/ValidationExceptionHandler.java
@@ -1,6 +1,6 @@
 package QCraft.QCraft.Handler;
 
-import QCraft.QCraft.dto.response.ResponseDTO;
+import QCraft.QCraft.dto.response.member.ResponseDTO;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.MethodArgumentNotValidException;
@@ -12,7 +12,6 @@ public class ValidationExceptionHandler {
 
     @ExceptionHandler({MethodArgumentNotValidException.class, HttpMessageNotReadableException.class})
     public ResponseEntity<ResponseDTO> validationExceptionHandler(Exception e) {
-        System.out.println("3");
         return ResponseDTO.validationError();
     }
 }
