+++ src/main/java/QCraft/QCraft/dto/request/member/CheckCertificationRequestDTO.java
@@ -1,4 +1,4 @@
-package QCraft.QCraft.dto.request;
+package QCraft.QCraft.dto.request.member;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -9,8 +9,13 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class EmailCertificationRequestDTO {
+public class CheckCertificationRequestDTO {
+
     @Email
     @NotBlank
     private String email;
+
+    @NotBlank
+    private String certificationNumber;
+
 }
