+++ src/main/java/QCraft/QCraft/dto/request/member/EmailCheckRequestDTO.java
@@ -1,6 +1,5 @@
-package QCraft.QCraft.dto.request;
+package QCraft.QCraft.dto.request.member;
 
-import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,12 +8,8 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class SignInRequestDTO {
+public class EmailCheckRequestDTO {
 
-    @Email
     @NotBlank
     private String email;
-
-    @NotBlank
-    private String password;
 }
