+++ src/main/java/QCraft/QCraft/dto/request/member/SignInRequestDTO.java
@@ -1,5 +1,6 @@
-package QCraft.QCraft.dto.request;
+package QCraft.QCraft.dto.request.member;
 
+import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,8 +9,12 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class EmailCheckRequestDTO {
+public class SignInRequestDTO {
 
+    @Email
     @NotBlank
     private String email;
+
+    @NotBlank
+    private String password;
 }
