+++ src/main/java/QCraft/QCraft/dto/request/member/SignUpRequestDTO.java
@@ -1,8 +1,7 @@
-package QCraft.QCraft.dto.request;
+package QCraft.QCraft.dto.request.member;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
