+++ src/main/java/QCraft/QCraft/dto/response/member/ResponseDTO.java
@@ -1,4 +1,4 @@
-package QCraft.QCraft.dto.response;
+package QCraft.QCraft.dto.response.member;
 
 import QCraft.QCraft.commons.ResponseCode;
 import QCraft.QCraft.commons.ResponseMessage;
